+++ Network/Wai/Handler/Warp.hs
@@ -124,7 +124,7 @@ module Network.Wai.Handler.Warp (
   , defaultPushPromise
   ) where
 
-import Control.Exception (SomeException, throwIO)
+import UnliftIO.Exception (SomeException, throwIO)
 import Data.Streaming.Network (HostPreference)
 import qualified Data.Vault.Lazy as Vault
 import qualified Network.HTTP.Types as H